# Add shareindexes macro for (sometimes) faster iteration