@eschnett 2016-03-04
It seems as if this could be solved by "control inversion": Instead of passing the iterator to the for loop, one could pass the for loop body (as closure) to the iterator, which can then easily implement the `if` statement outside the loop. If closures are fast (are they?), this could be an option. The front-end would need to lower `for` loops to a call to `Base.forloop`.

I know this sounds radical and far-fetched, and I'm only mentioning this tongue-in-cheek. However, it would be a quite natural solution to implementing parallel loops, (partial) unrolling, etc. Of course, a macro has the same expressive power, and it's thus safer (for guaranteed performance) to use a macro than to rely on loop hoisting or common subexpression elimination.