@mbauman 2016-03-04
That'd definitely be worth a test, @eschnett.  It doesn't need a complete front-end re-write:

```jl
eachindex(A, B, …) do (IA, IB, …)
    # body
end
```

Brave new world, indeed!