@timholy 2016-03-04
Great suggestion, @eschnett. I will test this (if @mbauman doesn't beat me to it). As @mbauman says, I suspect control over inlining may be necessary. I did notice that
```
function foo(A)
    @inline function inner(i)
        body
    end
    loop
end
```
is valid syntax. However, I am also noticing that setting external variables from a closure has inference problems. That is,
```jl
getith(i) -> A[i]
```
is fast, but
```jl
addith(i) -> (@inbounds s+= A[i])
```
is slow.