@tkelman 2016-03-05
also ref #12902