@nalimilan 2016-03-05
I don't understand why `eachindex(A, B)` is too limited. Couldn't it return a wrapper iterator (like `zip` does) that does whatever is needed, with a fast path for when a single index can be used for both arrays? I must say I don't really like the idea that standard `for` loops couldn't be used for efficient generic array indexing. That's one of their main uses!