@timholy 2016-03-05
It wasn't very clear, but what I meant is that having `eachindex` return a *single* index that is magically supposed to be good for both arrays is too limited.