@nalimilan 2016-03-05
Got it. Could we imagine writing this instead?
```julia
for (IA, IB) in eachindex(A, B)
    # body
end
```
That would allow the iterator to return twice the same index if possible.

(Yes, this would be a breaking change, but we could find another name for this function.)