@timholy 2016-03-05
Yes, that's exactly what the `zip(eachindex(A), eachindex(B))` does. I think I'd rather keep the name `eachindex`, and go through a release cycle where we deprecate the single-index version. Then for the release after that we can introduce the multi-index return.