@nalimilan 2016-03-05
The point is that `for (IA, IB) in eachindex(A, B)` could return a single index and increment it only once if possible, while `for (IA, IB) in zip(eachindex(A), eachindex(B))` cannot. That's what you're trying to fix, right?