@timholy 2016-03-05
Ideally yes, but as you wrote it that falls under the "An alternative that didn't work" section above.