@nalimilan 2016-03-05
Sorry, I may be missing something, but here's an illustration of what I have in mind for the 2-arrays case. The equality check would only happen once.
```julia
immutable PseudoZip2{I} <: Base.AbstractZipIterator
    a::I
end
Base.length(z::PseudoZip2) = length(z.a)
Base.eltype{I}(::Type{PseudoZip2{I}}) = Tuple{eltype(I), eltype(I)}
Base.start(z::PseudoZip2) = (s = start(z.a); (s, s))
function Base.next(z::PseudoZip2, st)
    n = next(z.a, st[1])
    return ((n[1], n[1]), (n[2], n[2]))
end
Base.done(z::PseudoZip2, st) = done(z.a, st[1])

function eachindex_shared(A, B)
    RA = eachindex(A)
    RB = eachindex(B)

    if RA == RB
        PseudoZip2(eachindex(A))
    else
        zip(eachindex(A), eachindex(B))
    end
end

for I(A, IB) in eachindex_shared(A, B)
    # ...
end
```