@timholy 2016-03-05
Yes, I'm trying to avoid the type-instability. Worse, you can't do this purely on the basis of the types of `RA` and `RB`. For example, say you're copying the elements of a 4x6 array to a 6x4 array. The two iterators are not equal (so you'd want to use regular `Zip2`), but they differ only in the value domain, not in the type domain.