@toivoh 2016-03-05
If you could design the iterator to make llvm perform loop unswitching (use
conditional code in the loop that obviously depends on a loop invariant
condition) then it might be possible to get llvm to duplicate the loop for
us and choose which one based on the value domain information at runtime.
But I'm not sure how robust that would be, even if it could be made to
work.