@timholy 2016-03-05
I basically tried that: that `Zip2Each` structure in the top post is `immutable`, and I think I tried to force-inline `start`, `next`, and `done`. Maybe I screwed up want to give it a try yourself and see if you can pull it off?