@nalimilan 2016-03-05
I wondered whether, for `zip(eachindex(A), eachindex(B))`, LLVM shouldn't be able to detect that the code is incrementing two indices in parallel and that they cannot possibly be different, so that only one of them needs to be computed. Indeed, if you change:
```julia
Base.done(z::Base.Zip2, st) = done(z.a,st[1]) | done(z.b,st[2])
```
to
```julia
Base.done(z::Base.Zip2, st) = done(z.a,st[1])
```
then the generated code for the loop is identical to that obtained with `enumerate(A, B)` when passing two `Arrays` of the same size.

This kind of optimization could be applied via a special, say, `SameLengthZip2` type which would check for equal lengths of its inputs. I guess `eachindex` isn't intended to be used with arrays of different lengths, right? Maybe LLVM could also be improved about this particular case, since it's already smart enough to get rid of the double computation.

Code to replicate:
```julia
A = rand(10000, 1000);
B = rand(10000, 1000);

function f(A, B)
    s = 0.0
    for I in eachindex(A, B)
        @inbounds s += A[I]*B[I]
    end
    s
end

Base.done(z::Base.Zip2, st) = done(z.a,st[1])

function g(A, B)
    s = 0.0
    for (IA, IB) in zip(eachindex(A), eachindex(B))
        @inbounds s += A[IA]*B[IB]
    end
    s
end

@code_native f(A, B)
@code_native g(A, B)
```