@timholy 2016-03-05
@nalimilan, that looks really promising! I'll check that myself (or feel free to submit your own PR, since you discovered this).

We'd probably need a separate function, since `zip` is designed to handle mismatched iterators. `zipmatched`? Or `zip(Val{:matched}, iter1, iter2, ...)`?