@timholy 2016-03-05
I'm not replicating this, @nalimilan, and the case of two `LinearSlow` arrays (`AS = sub(A, 1:size(A,1), :)`) is still much slower. Can you check to see if it works for you?