@nalimilan 2016-03-06
As you found out, it only works for `LinearFast`, since for `LinearSlow` the code needs to handle the possibility of arrays with different shapes, e.g. `(2, 3)` and `(3, 2)`. That said, I'm surprised you don't see the same efficient code as I do for `LinearFast`. Also, this means LLVM isn't that far from being able to optimize this itself.

Could one of our LLVM experts comment here? It looks like LLVM should be able to do this, at least if the correct passes are enabled in the expected order. There's also the issue of the threshold. http://llvm.org/docs/Passes.html#loop-unswitch-unswitch-loops

I've written a small C program to see whether clang is able to hoist the check for `same` in a similar loop as ours: https://gist.github.com/nalimilan/fffd490da8686ef72ab2 From my (limited) understanding of the assembly output (`clang -S -masm=intel unswitch.c`), it looks like the hoisting only happens with `-O3`, not with `-O2`. So maybe we only need to tweak the thresholds, or have a way to tell the compiler what it has to do in this particular case. It would be interesting to see what happens with a more complex example similar to `LinearSlow`, but that's a bit more work.